Try to workaround this stupid NPE
It seems android tries to restore those !#$#& fragments before
the activity has finished creating itself. So their oncreateview
goes to hell. Fuck these fragments 100 times over, then quarter
them with horses and then stomp on the pieces and feed them to worms.

In fact the problem is all because of the way Services are defined.
There is no synchronous way to bind to a service, even a local one.
Because of this, there is no valid way to write what I want to do.

Android services suck. The next time around I will NOT use them.
Except for EmptyService, like Contacts uses. That's all services are
good for.